refactor: modularize HTML parser to new HtmlParser file
