+++ compose-highlight/src/main/kotlin/dev/hossain/highlight/engine/internal/HtmlToAnnotatedString.kt
@@ -10,18 +10,6 @@ import dev.hossain.highlight.engine.HljsSelectors
 import kotlin.time.Duration
 import kotlin.time.measureTimedValue
 
-internal sealed interface CustomNode
-
-internal class CustomElement(
-    val tagName: String,
-    val className: String,
-    val childNodes: List<CustomNode>,
-) : CustomNode
-
-internal class CustomTextNode(
-    val wholeText: String,
-) : CustomNode
-
 /**
  * Converts Highlight.js HTML output into a Compose [AnnotatedString].
  *
@@ -288,230 +276,6 @@ internal object HtmlToAnnotatedString {
         // Fall back to the first recognized class.
         return classes.firstNotNullOfOrNull { colorMap[it] }
     }
-}
-
-/**
- * Parses a simple HTML fragment into a lightweight CustomNode tree.
- * Supports basic elements (like span), attributes (like class), HTML comments,
- * and standard HTML entity decoding.
- */
-internal fun parseHtml(html: String): List<CustomNode> {
-    var index = 0
-    val length = html.length
-
-    fun peek(offset: Int = 0): Char? {
-        val i = index + offset
-        return if (i < length) html[i] else null
-    }
-
-    fun decodeEntities(text: String): String {
-        if (!text.contains('&')) return text
-        val sb = StringBuilder(text.length)
-        var i = 0
-        val len = text.length
-        while (i < len) {
-            val c = text[i]
-            if (c == '&') {
-                val semi = text.indexOf(';', i)
-                if (semi != -1 && semi - i < 10) {
-                    val entity = text.substring(i + 1, semi)
-                    when (entity) {
-                        "amp" -> {
-                            sb.append('&')
-                        }
-
-                        "lt" -> {
-                            sb.append('<')
-                        }
-
-                        "gt" -> {
-                            sb.append('>')
-                        }
-
-                        "quot" -> {
-                            sb.append('"')
-                        }
-
-                        "apos" -> {
-                            sb.append('\'')
-                        }
-
-                        "nbsp" -> {
-                            sb.append('\u00A0')
-                        }
-
-                        else -> {
-                            if (entity.startsWith("#")) {
-                                val code =
-                                    if (entity.startsWith("#x")) {
-                                        entity.substring(2).toIntOrNull(16)
-                                    } else {
-                                        entity.substring(1).toIntOrNull(10)
-                                    }
-                                if (code != null) {
-                                    sb.append(code.toChar())
-                                } else {
-                                    sb.append('&').append(entity).append(';')
-                                }
-                            } else {
-                                sb.append('&').append(entity).append(';')
-                            }
-                        }
-                    }
-                    i = semi + 1
-                } else {
-                    sb.append('&')
-                    i++
-                }
-            } else {
-                sb.append(c)
-                i++
-            }
-        }
-        return sb.toString()
-    }
-
-    fun extractClassAttr(attrsString: String): String {
-        var i = 0
-        val len = attrsString.length
-        while (i < len) {
-            while (i < len && attrsString[i].isWhitespace()) {
-                i++
-            }
-            if (i >= len) break
-
-            val eq = attrsString.indexOf('=', i)
-            if (eq == -1) break
-            val attrName = attrsString.substring(i, eq).trim().lowercase()
-            i = eq + 1
-
-            while (i < len && attrsString[i].isWhitespace()) {
-                i++
-            }
-            if (i >= len) break
-
-            val quote = attrsString[i]
-            if (quote == '"' || quote == '\'') {
-                i++
-                val endQuote = attrsString.indexOf(quote, i)
-                if (endQuote != -1) {
-                    val attrValue = attrsString.substring(i, endQuote)
-                    if (attrName == "class") {
-                        return attrValue
-                    }
-                    i = endQuote + 1
-                } else {
-                    val attrValue = attrsString.substring(i)
-                    if (attrName == "class") {
-                        return attrValue
-                    }
-                    i = len
-                }
-            } else {
-                var endValue = i
-                while (endValue < len && !attrsString[endValue].isWhitespace()) {
-                    endValue++
-                }
-                val attrValue = attrsString.substring(i, endValue)
-                if (attrName == "class") {
-                    return attrValue
-                }
-                i = endValue
-            }
-        }
-        return ""
-    }
-
-    fun parseNodes(parentTag: String?): List<CustomNode> {
-        val nodes = mutableListOf<CustomNode>()
-        while (index < length) {
-            val c = peek() ?: break
-            if (c == '<') {
-                if (html.startsWith("<!--", index)) {
-                    val endComment = html.indexOf("-->", index + 4)
-                    if (endComment != -1) {
-                        index = endComment + 3
-                    } else {
-                        index = length
-                    }
-                    continue
-                }
-
-                if (peek(1) == '/') {
-                    val tagEnd = html.indexOf('>', index + 2)
-                    val tagName =
-                        if (tagEnd != -1) {
-                            html.substring(index + 2, tagEnd).trim().lowercase()
-                        } else {
-                            ""
-                        }
-                    if (tagEnd != -1 && tagName == parentTag) {
-                        index = tagEnd + 1
-                        break
-                    } else if (tagEnd != -1) {
-                        index = tagEnd + 1
-                        continue
-                    } else {
-                        index = length
-                        break
-                    }
-                }
-
-                val tagEnd = html.indexOf('>', index + 1)
-                if (tagEnd == -1) {
-                    val text = html.substring(index)
-                    nodes.add(CustomTextNode(decodeEntities(text)))
-                    index = length
-                    break
-                }
-
-                val tagContent = html.substring(index + 1, tagEnd).trim()
-                val isSelfClosing = tagContent.endsWith('/')
-                val cleanContent = if (isSelfClosing) tagContent.dropLast(1).trim() else tagContent
-
-                val firstSpace = cleanContent.indexOfAny(charArrayOf(' ', '\t', '\r', '\n'))
-                val tagName =
-                    if (firstSpace != -1) {
-                        cleanContent.substring(0, firstSpace).lowercase()
-                    } else {
-                        cleanContent.lowercase()
-                    }
-
-                var className = ""
-                if (firstSpace != -1) {
-                    val attrsString = cleanContent.substring(firstSpace + 1)
-                    className = extractClassAttr(attrsString)
-                }
-
-                index = tagEnd + 1
-
-                if (isSelfClosing) {
-                    nodes.add(CustomElement(tagName, className, emptyList()))
-                } else {
-                    val children = parseNodes(tagName)
-                    nodes.add(CustomElement(tagName, className, children))
-                }
-            } else {
-                val nextTag = html.indexOf('<', index)
-                val text =
-                    if (nextTag != -1) {
-                        val t = html.substring(index, nextTag)
-                        index = nextTag
-                        t
-                    } else {
-                        val t = html.substring(index)
-                        index = length
-                        t
-                    }
-                if (text.isNotEmpty()) {
-                    nodes.add(CustomTextNode(decodeEntities(text)))
-                }
-            }
-        }
-        return nodes
-    }
-
-    return parseNodes(null)
 }
 
 /**
